Rest api is done.

+++ dao/src/main/resources/create-db.sql
@@ -13,10 +13,10 @@ CREATE TABLE title
 (
     title_id      INT         NOT NULL AUTO_INCREMENT,
     name          VARCHAR(50) NOT NULL,
-    budget        LONG        NOT NULL,
+    budget        REAL        NOT NULL,
     premiere_date DATE        NOT NULL,
     runtime       TIME        NOT NULL,
-    box_office    LONG        NOT NULL,
+    box_office    REAL        NOT NULL,
     director_id   INT,
     CONSTRAINT fk_director_id FOREIGN KEY (director_id)
         REFERENCES director (director_id)
